+++ TelegramBotAPI-core/src/commonMain/kotlin/com/github/insanusmokrassar/TelegramBotAPI/bot/Ktor/base/MultipartRequestCallFactory.kt
@@ -1,6 +1,7 @@
 package com.github.insanusmokrassar.TelegramBotAPI.bot.Ktor.base
 
 import com.github.insanusmokrassar.TelegramBotAPI.requests.abstracts.*
+import com.github.insanusmokrassar.TelegramBotAPI.utils.TelegramAPIUrlsKeeper
 import com.github.insanusmokrassar.TelegramBotAPI.utils.mapWithCommonValues
 import io.ktor.client.HttpClient
 import io.ktor.client.request.forms.MultiPartFormDataContent
@@ -8,11 +9,10 @@ import io.ktor.client.request.forms.formData
 import io.ktor.http.Headers
 import io.ktor.http.HttpHeaders
 
-class MultipartRequestCallFactory : AbstractRequestCallFactory() {
-
+object MultipartRequestCallFactory : AbstractRequestCallFactory() {
     override fun <T : Any> prepareCallBody(
         client: HttpClient,
-        baseUrl: String,
+        urlsKeeper: TelegramAPIUrlsKeeper,
         request: Request<T>
     ): Any? = (request as? MultipartRequest) ?.let { castedRequest ->
         MultiPartFormDataContent(