+++ TelegramBotAPI-extensions-utils/src/commonMain/kotlin/com/github/insanusmokrassar/TelegramBotAPI/extensions/utils/shortcuts/MediaGroupsShortcuts.kt
@@ -4,15 +4,15 @@ import com.github.insanusmokrassar.TelegramBotAPI.requests.send.media.SendMediaG
 import com.github.insanusmokrassar.TelegramBotAPI.types.*
 import com.github.insanusmokrassar.TelegramBotAPI.types.chat.abstracts.Chat
 import com.github.insanusmokrassar.TelegramBotAPI.types.message.ForwardInfo
-import com.github.insanusmokrassar.TelegramBotAPI.types.message.abstracts.MediaGroupMessage
-import com.github.insanusmokrassar.TelegramBotAPI.types.message.abstracts.Message
+import com.github.insanusmokrassar.TelegramBotAPI.types.message.abstracts.*
+import com.github.insanusmokrassar.TelegramBotAPI.types.message.content.abstracts.MediaGroupContent
 import com.github.insanusmokrassar.TelegramBotAPI.types.update.MediaGroupUpdates.SentMediaGroupUpdate
 
-val List<MediaGroupMessage>.forwardInfo: ForwardInfo?
+val List<CommonMessage<MediaGroupContent>>.forwardInfo: ForwardInfo?
     get() = firstOrNull() ?.forwardInfo
-val List<MediaGroupMessage>.replyTo: Message?
+val List<CommonMessage<MediaGroupContent>>.replyTo: Message?
     get() = firstOrNull() ?.replyTo
-val List<MediaGroupMessage>.chat: Chat?
+val List<CommonMessage<MediaGroupContent>>.chat: Chat?
     get() = firstOrNull() ?.chat
 val List<MediaGroupMessage>.mediaGroupId: MediaGroupIdentifier?
     get() = firstOrNull() ?.mediaGroupId
@@ -26,7 +26,7 @@ val SentMediaGroupUpdate.chat: Chat
 val SentMediaGroupUpdate.mediaGroupId: MediaGroupIdentifier
     get() = data.mediaGroupId!!
 
-fun List<MediaGroupMessage>.createResend(
+fun List<CommonMessage<MediaGroupContent>>.createResend(
     chatId: ChatId,
     disableNotification: Boolean = false,
     replyTo: MessageIdentifier? = null
@@ -37,7 +37,7 @@ fun List<MediaGroupMessage>.createResend(
     replyTo
 )
 
-fun List<MediaGroupMessage>.createResend(
+fun List<CommonMessage<MediaGroupContent>>.createResend(
     chat: Chat,
     disableNotification: Boolean = false,
     replyTo: MessageIdentifier? = null